Removed SSLv2Hello protocol for http client.
Otherwise the SSLEngine will automatically use since it's the protocol
with the least version but then SNI won't work. Also there's no
benefit in enabling it since whenever the server does not support
TLS1.0+ the client won't connect anyway, so in those cases just a
different exception should occur which will be wrapped by HttpClient
in the same way: as a SSLHandshakeException without a further cause.

Signed-off-by: Jörg Eichhorn <joerg.eichhorn@kiwigrid.com>